+++ common/src/main/java/io/coti/common/data/TransactionIndexData.java
@@ -3,17 +3,24 @@ package io.coti.common.data;
 import io.coti.common.data.interfaces.IEntity;
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class TransactionIndexData implements IEntity {
 
     private transient Hash hash;
-    private long index;
+    private BigInteger index;
 
     private TransactionIndexData() {
     }
 
-    public TransactionIndexData(Hash hash, long index) {
+    public TransactionIndexData(Hash hash, BigInteger index) {
         this.hash = hash;
         this.index = index;
     }
+
+
+
+
+
 }
